+++ src/main/java/com/didacto/controller/v1/pay/WebhookController.java
@@ -4,9 +4,12 @@ package com.didacto.controller.v1.pay;
 import com.didacto.dto.pay.WebhookPayloadRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+
 @RestController
+@RequestMapping("/api/v1")
 public class WebhookController {
 
     @PostMapping("/webhook")
