+++ src/main/java/com/didacto/domain/Member.java
@@ -29,8 +29,6 @@ public class Member extends BaseEntity {
 
     @Column(nullable = false)
     private String name;
-
-
     @Column
     private OffsetDateTime birth;
