+++ src/main/java/com/didacto/domain/Order.java
@@ -1,6 +1,7 @@
 package com.didacto.domain;
 
 
+import com.didacto.common.BaseEntity;
 import com.didacto.dto.pay.WebhookPayloadRequest;
 import jakarta.persistence.*;
 import lombok.Builder;
@@ -11,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @Table(name = "orders") // Mariadb에서 order가 예약어로 되어있기 때문에 @Table은 필수로 작성
-public class Order {
+public class Order extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
